# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEC",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/",
            "description": "DEV"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Listar todos os usuários",
                "description": "Lista todos os usuários: 3 tipos de usuários obtem o acesso desse endpoint: ADMINISTRADOR, REPRESENTANTE E GERENTE",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "nome do email que deseja filtrar",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Lista o registro de usuários por ID",
                "description": "Lista o registro de usuários por ID de referência",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Usuário not found"
                    },
                    "200": {
                        "description": "Ok"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Atualizar usuário",
                "description": "Atualizar usuário: Apenas o usuário pode atualizar suas próprias informações",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "name": "Nome do usuário",
                                    "email": "Email do usuário",
                                    "password": "Senha do usuário"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Usuário not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Deletar usuário",
                "description": "Deletar usuário por ID de referência: Apenas o usuário pode deletar suas próprias informações ou o administrador",
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Usuario Not Found"
                    }
                }
            }
        },
        "/users/set-user/{id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Atualiza usuário para se tornar gerente ou volta usuário para visualizador",
                "description": "Transforma usuário em gerente por ID de referência: Apenas o administrador tem acesso",
                "operationId": "b1e66f3e1f8d90059556aca8f6f7246c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "isManager": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Usuario Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ],
...
}